Move DisplayQuestion to be shared
