+++ src/main/js/screens/app/shared/DisplayQuestion.js
